requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
cached-property==2.0.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
crypto==1.4.1
ecdsa==0.19.0
fastapi==0.115.6
fastapi-auth0==0.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
Mako==1.3.8
MarkupSafe==3.0.2
mypy-extensions==1.0.0
Naked==0.1.32
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
shellescape==3.8.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.36
sqlalchemy-orm==1.2.10
starlette==0.41.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
